[gdb/cli] Allow source highlighting to be interrupted (continued)
In PR cli/30934, a problem is reported where gdb becomes unresponsive for
2m13s because the GNU source-highlight library is taking a lot of time to
annotate the source:
...
$ gdb -q a.out -ex "b 56"
Reading symbols from a.out...
Breakpoint 1 at 0x400e2a: file test.cpp, line 56.
(gdb) run
Starting program: /data/vries/gdb/a.out

Breakpoint 1, Solution::numOfSubarrays () at test.cpp:56
<... wait for it ...>
56	        return (int) t;
(gdb)
...

This is due to a problem in the library [1], for which I've posted a
patch [2], which brings down the annotation time to 3s.

However, GDB should not become unresponsive due to such a problem.

The previous patch allows us to interrupt a list command during highlighting
by either SIGTERM or SIGINT.

In this case, we can use SIGTERM to interrupt the run command, but not SIGINT.

This is due to the fact that the infrun_quit_handler is still active.

The purpose of infrun_quit_handler is to:
- if !target_terminal::is_ours, pass ^C to the inferior, and
- if target_terminal::is_ours, ignore ^C.  If gdb is executing say a continue
  in the background, it generates events and we don't want to interrupt the
  handling of these events.

In this case however, the reason target_terminal::is_ours is not because
there's background execution, but because normal_stop has decided that we have
stopped execution, and it's preparing to hand back control to GDB.

Fix this by switching back to the default_quit_handler during notify_normal_stop.

Tested on x86_64-linux.

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=30934

[1] https://savannah.gnu.org/bugs/?64747
[2] https://savannah.gnu.org/patch/index.php?10400