[gdb/cli] Allow source-highlight to autodetect language
Currently when gdb asks the source-highlight library to highlight a file, it
tells it what language file to use.

However, if it doesn't know what language file to use, it falls back to using
python pygments, and in absence of that, unhighlighted source text.

In the case of python pygments, it autodetects which language to use based on
the file name.

Add the same capability when using the source-highlight library.

Tested on x86_64-linux.

Verified that it works by making get_language_name return nullptr for
language_c, and verifying that source-highlight still manages to hightlight a
hello world.

PR cli/30966
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=30966