[gdb] Detect misplaced AT_NULL in target_auxv_search
When starting to investigate PR30547, I noticed that I only ran into it on
ppc64 and s390x, two big-endian architectures, so I sort of expected to find
an endian-related problem.

Instead, the problem was ppc_linux_target_wordsize returning 4 instead of 8.

The same problem happens on ppc64le (ppc_linux_target_wordsize returns 4),
it's just that the incorrect word size doesn't change the outcome of:
- target_auxv_search (AT_NULL, &dummy) == 1, and
- target_auxv_search (AT_PAGESZ, &dummy) == 1
so linux_is_uclinux returns false, as it should.

This suggest a too forgiving parsing of the auxv vector.

Fix this by requiring in target_auxv_search that AT_NULL, if found, comes at
the end of the auxv vector.

This results in target_auxv_search (AT_NULL, &dummy) == -1, which still means
that linux_is_uclinux returns false, and PR30547 doesn't reproduce.

Tested on x86_64-linux.