[gdb] Add asserts in linux_is_uclinux
In function linux_is_uclinux:
...
  return (target_auxv_search (AT_NULL, &dummy) > 0
	  && target_auxv_search (AT_PAGESZ, &dummy) == 0);
...
if target_auxv_search returns -1, it means the auxv vector is incorrect, and
we can't really decide whether we have uclinux or not.

Add asserts after each call to detect this situation.

This allows to reproduce PR30547 on ppc64le (if we indeed revert the fix).

Tested on x86_64-linux.