[gdb/dap] Fix race between dap exit and gdb exit
When DAP shuts down due to an EOF event, there's a race between:
- gdb's main thread handling a SIGHUP, and
- the DAP main thread exiting.

Fix this by waiting for DAP's main thread exit during the gdb_exiting event.

Because the _dap_thread is set in the DAP main thread, it's not available in
gdb's main thread, so we have to pass it there using a queue.

I've also tried moving the setting of _dap_thread to gdb's main thread to
prevent needing to use queue, but I didn't manage to make that work.

Tested on aarch64-linux.

PR dap/31380
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=31380