[gdb] Don't init history in batch mode
I noticed in captured_main_1 that init_history is called before bailing out
if batch_flag is true.

The history is used only in an interactive session, so there's no need to
initialize it in batch mode.

Fix this by only doing init_history if we've established we're not in batch
mode.

Tested on aarch64-linux.