[gdb/dap] Don't let dap log string grow indefinitely
As mentioned in commit "[gdb/dap] Fix race between dap startup and dap log
file", initially logging to a string has the drawback that the string can keep
growing indefinitely if a log file is never set.

Fix this using the gdb_starting_session event to detect this situation, and
stopping the logging to string.

Tested on aarch64-linux.

PR dap/31386
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=31386