[gdb] Add gdb_starting_session observable
Add a new observable gdb_starting_session, that triggers when gdb has finished
processing the command line options, and is starting an interactive session.

This is not triggered in batch mode.

The observable doesn't have any users yets, but a user will be added in
a following patch.

Tested on aarch64-linux.