[gdb/dap] Fix stray KeyboardInterrupt after cancel
When running test-case gdb.dap/pause.exp 100 times in a loop, it passes
100/100.

But if we remove the two "sleep 0.2" from the test-case, we run into
(copied from dap.log and edited for readability):
...
Traceback (most recent call last):
  File "startup.py", line 251, in message
    def message():

KeyboardInterrupt
Quit
...

This happens as follows.

CancellationHandler.cancel calls gdb.interrupt to cancel a request in flight.

The idea is that this interrupt triggers while in fn here in message (a nested
function of send_gdb_with_response):
...
    def message():
        try:
            val = fn()
            result_q.put(val)
        except (Exception, KeyboardInterrupt) as e:
            result_q.put(e)
...
but instead it triggers outside the try/except.

Fix this by:
- in CancellationHandler, adding a variable interruptable to track whether
  we're in the try/except,
- in CancellationHandler, adding a new contextmanager interruptable_region
  that sets variable interruptable to True,
- in message, wrapping "val = fn()" in the interruptable_region
- in CancellationHandler.cancel, if interruptable == False, setting a variable
  cancel_pending to True instead of callling gdb.interrupt , and
- bailing out of interruptable_region at entry if cancel_pending == True.

This makes the test-case pass 100/100, also when adding the extra stressor of
"taskset -c 0", which makes the fail more likely without the patch.

Note that the fix means that we no longer attempt to interrupt a request without
result (a request that is handled using send_gdb rather than
send_gdb_with_reponse).  Such request can still be cancelled, just not when
"in flight".

For a request without result, the "in flight" period corresponds to the
queuing of the request on the runnables queue, and consequently calling
gdb.interrupt during this window may result in the cancellation of another
request.  To allow cancellation of this type of request after queuing in the
runnables queue, a more elaborate approach is needed.  The focus of this
commit is merely to fix behaviour in the current setup.

Tested on aarch64-linux.

PR dap/31275
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=31275