[binutils/readelf] Add fixme in display_debug_str_offsets
While reading display_debug_str_offsets, I realized that the v4 split dwarf
handling bit doesn't assign entry_length, so it can be either 4 (32-bit dwarf)
or 8 (64-bit dwarf).  However, I think the cases for which it'll be 8 are
rare and odd enough to conclude that in practise we default to 32-bit
dwarf.

As mentioned in the fission documentation [1], whether a .debug_str_offsets
section is 32-bit or 64-bit dwarf depends on the compilation unit header in
the .debug_info.dwo section, which is currently not inspected.

For now:
- make an explicit choice for 32-bit dwarf by assiging 4 to entry_length, and
- add a fixme comment.

[1] https://gcc.gnu.org/wiki/DebugFission